remove xfail from tests and move ensure_df_mutable to _arrow module
